Move nfpm fields applicable to all targets to nfpm.fields.all
